Set up for testing
Added junit and maven plugins for code coverage and run junit tests.
Added junit test classes for each class.
Updated package structure to better represent maven project.
