+++ README.md
@@ -1,2 +1,2 @@
 # ImmutableList
-Immutable list is known as a list that can only be accessed. In this repository, is a simple Java data structure implementation for ImmutableList that supports mutation method by creating a new reference.
+An immutable list is known as a list that can only be accessed. In this repository, is a few Java data structure implementations for ImmutableList that supports mutation methods by creating a new list that represents the new state after mutation.
