+++ src/main/java/com/lumingwu/immutable/list/Stack.java
@@ -10,7 +10,7 @@
  * and limitations under the License.
  */
 
-package main.java.com.lumingwu;
+package com.lumingwu.immutable.list;
 
 /**
  * Stack -  A simple dedicated singly-linked Stack for ImmutableArrayList search operation.
