+++ src/test/java/com/lumingwu/immutable/list/ImmutableArrayListTest.java
@@ -1,0 +1,19 @@
+/*
+ * Copyright 2020 Luming Wu. All Rights Reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License"). You may not use this file except in compliance with the License. A copy of the License is located at
+ *
+ * https://github.com/LumingWu/ImmutableList/blob/master/LICENSE
+ *
+ * or in the "license" file accompanying this file. This file is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR
+ * CONDITIONS OF ANY KIND, either express or implied. See the License for the specific language governing permissions
+ * and limitations under the License.
+ */
+
+package com.lumingwu.immutable.list;
+
+public class ImmutableArrayListTest {
+
+
+
+}
